refactoring(03): decompose ResultsPage.tsx into container + sections + editor
Split the 3177-line ResultsPage.tsx into:
- ResultsPage.tsx (~1857): container only (byte-identical body)
- resultSections.tsx (~1161): presentational subcomponents + pure helpers + types
- PublishYamlEditor.tsx (~162): the publish.yaml (EAR allowlist) editor

Code moved verbatim; only export keywords and import headers added. Verified
byte-for-byte against HEAD; adversarial 2-lens review confirmed behavior
preserved (component identity stable, no logic/JSX/ordering change).

Checks: typecheck 0 non-test errors (11 pre-existing __tests__ errors only);
build ok; vitest 4 passed/2 failed (pre-existing brittle tests). Remaining large
components + finer splits + container seams moved to follow-up req 15. Analysis
recorded in refactoring/notes/03_resultspage_decomposition.md. Req 03 file
deleted; completion recorded in COMPLETED.md.
