+++ ari-core/ari/viz/frontend/src/README.md
@@ -57,7 +57,9 @@ React/TypeScript source for the ARI `ari.viz` web dashboard — app entry, pages
   - `Results/` — final run results and rubric scoring.
     - `README.md` — Results index.
     - `index.ts` — barrel re-export.
-    - `ResultsPage.tsx` — results page container.
+    - `PublishYamlEditor.tsx` — per-checkpoint publish.yaml (EAR allowlist) editor;
+    - `resultSections.tsx` — presentational subcomponents and pure helpers for the
+    - `ResultsPage.tsx` — results page container (state, data loading, layout).
     - `RubricTreeVisualization.tsx` — D3 rubric tree with aggregated leaf scores.
   - `Settings/` — dashboard/run configuration page.
     - `README.md` — Settings index.
