+++ ari-core/ari/viz/frontend/src/components/README.md
@@ -52,7 +52,9 @@ Feature-grouped React components — one subdirectory per dashboard page plus sh
 - `Results/` — final run results and rubric scoring.
   - `README.md` — Results index.
   - `index.ts` — barrel re-export.
-  - `ResultsPage.tsx` — results page container.
+  - `PublishYamlEditor.tsx` — per-checkpoint publish.yaml (EAR allowlist) editor;
+  - `resultSections.tsx` — presentational subcomponents and pure helpers for the
+  - `ResultsPage.tsx` — results page container (state, data loading, layout).
   - `RubricTreeVisualization.tsx` — D3 rubric tree with aggregated leaf scores.
 - `Settings/` — dashboard/run configuration page.
   - `README.md` — Settings index.
