+++ ari-core/ari/viz/frontend/src/components/Results/README.md
@@ -6,5 +6,7 @@ Results page — final run results and rubric scoring.
 
 - `README.md` — this file.
 - `index.ts` — barrel re-export.
-- `ResultsPage.tsx` — results page container.
+- `PublishYamlEditor.tsx` — per-checkpoint publish.yaml (EAR allowlist) editor;
+- `resultSections.tsx` — presentational subcomponents and pure helpers for the
+- `ResultsPage.tsx` — results page container (state, data loading, layout).
 - `RubricTreeVisualization.tsx` — D3 rubric tree with aggregated leaf scores.
